pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "telegram bot with admin panel"
authors = [
    {name = "Vitalyi",email = "evciu97@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "python-telegram-bot (>=21.10,<22.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
